% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gradientDecent}
\alias{gradientDecent}
\title{gradientDecent}
\usage{
gradientDecent(X.mat, y.vec, step.size, max.iterations)
}
\arguments{
\item{X.mat}{scaled matrix of features}

\item{y.vec}{labels to X.mat}

\item{step.size}{the step size for the gradient decent}

\item{max.iterations}{the max number of iterations to perform the decent}
}
\value{
a matrix of weights for each iteration [ncol(X.mat) : max.iterations]
}
\description{
gradientDecent
}
\examples{


}
