% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{k.fold.cv}
\alias{k.fold.cv}
\title{k.fold.cv}
\usage{
k.fold.cv(X.mat, y.vec, compute.predictions, fold.vec)
}
\arguments{
\item{X.mat}{scaled data matrix}

\item{y.vec}{labels for the data}

\item{compute.predictions}{a function to make the predictions}

\item{fold.vec}{a vector of length nrow(X.mat) telling which fold each observation belongs to}
}
\value{
returns a vector of errors for each fold length(return.vec) = n.folds
}
\description{
runs an algorithm with a fold vector to get get an array of errors for each fold number 1:k-folds
}
\examples{


}
