% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelL1.R
\name{LinearModelL1}
\alias{LinearModelL1}
\title{LinearModelL1}
\usage{
LinearModelL1(
  X.scaled.mat,
  y.vec,
  penalty,
  opt.thresh,
  initial.weight.vec,
  step.size
)
}
\arguments{
\item{X.scaled.mat}{the scaled training matrix [ n_observations : n_features ]}

\item{y.vec}{the label vector for the data matrix [ n_observations : 1 ]}

\item{penalty}{a positive scaler penalty term}

\item{opt.thresh}{a positive scaler}

\item{initial.weight.vec}{a starting weight vector, can be random numbers or 0s [ n_features : 1 ]}

\item{step.size}{the incrimental step to descent with}
}
\value{
an optimal weight vector for predictings [n_features + 1 : 1] where the first entry is the intercept term
}
\description{
minimizes the following cost function: 1/n ∑i=1^n L[w^T x_i + b, y_i] + penalty * ||w||_1
}
\examples{
# load the default dataset
getwd()


}
