% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearModelL1.R
\name{LinearModelL1penalties}
\alias{LinearModelL1penalties}
\title{LinearModelL1penalties}
\usage{
LinearModelL1penalties(X.mat, y.vec, penalty.vec, step.size)
}
\arguments{
\item{X.mat}{an unscaled matrix of [ n_observations : n_features ]}

\item{y.vec}{labels for the observations [ n_observations : 1 ]}

\item{penalty.vec}{a vector of possible penelty values to test [ n_penalties : 1 ]}

\item{step.size}{the incrimental size to descent by}
}
\value{
a matrix of weights for each feature for each penalty used [ n_features + 1 : n_penalties ] (first row is the bias)
}
\description{
train different models based on the different penalty values in the penalty_vec and create a matric of weights for all of the penalties
W.mat (n_features+1 x n_penalties), weight matrix on original scale can be used to make
predictions via cbind(1, X.mat) %*% W.mat (the first row of W.mat should be the bias/beta/intercept)
}
\examples{


}
