% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunarBand.R
\docType{data}
\name{lunarData}
\alias{lunarData}
\title{Lunar Data from LROC WAC}
\format{
A data frame with 4,183,770 rows and 10 features:
\describe{
  \item{DN}{surface scattering}
  \item{Phase}{The angle }
  \item{Emission}{DESCRIPTION}
  \item{Incidence}{DESCRIPTION}
  \item{LEmission}{DESCRIPTION}
  \item{LIncidence}{DESCRIPTION}
  \item{Lat}{DESCRIPTION}
  \item{Long}{DESCRIPTION}
  \item{SunAz}{DESCRIPTION}
  \item{CraftAz}{DESCRIPTION}
}
}
\source{
\url{http://wms.lroc.asu.edu/lroc}
}
\usage{
lunarData
}
\description{
A dataset containing the DN, phase angle, emission angle, incidence angle, local emission angle, local incidence angle, latitude,
 longitude, Sun azimuthal direction and scapecraft azimuthal direction.
}
\keyword{datasets}
