% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_sq_error}
\alias{mean_sq_error}
\title{mean.sq.error}
\usage{
mean_sq_error(actual, predicted)
}
\arguments{
\item{actual}{the real values}

\item{predicted}{the values that were predicted}
}
\value{
returns an error value between 0 and 1
}
\description{
calculates the mean squared error of the prediction vector
}
\examples{


}
