% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MSE}
\alias{MSE}
\title{mean.sq.error}
\usage{
MSE(actual, predicted)
}
\arguments{
\item{actual}{vector of real labels}

\item{predicted}{vector of predicted labels}
}
\value{
the MSE of the prediction
}
\description{
this function finds the MSE between two vectors
}
\examples{

}
