% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{The object to apply sigmoid to.
Types can be: Numeric, Double, Integer, Vector, Data Frame}
}
\value{
returns value is of type x
}
\description{
Calculate the sigmoid for data given by x
}
\examples{
# Numerical
# ---------
val <- 25.6548
sigmoid(val)

# Vector
# ------
vector <- c(9.6574, 13.52, 22.34)
sigmoid(vector)

# Data Frame
# ----------
df <- data.frame(column_1=c(9.6574, 3.52, 232.34), column_2=c(9.6574, 13.52, 22.43))
sigmoid(df)

}
