% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minnaert}
\alias{minnaert}
\title{minnaert}
\usage{
minnaert(emission, incidence, k)
}
\arguments{
\item{emission}{the emission angle of the observation}

\item{incidence}{the incidence angle of the observation}

\item{k}{constant between 0 and 1}
}
\value{
returns minnaert prediction based on emission and incidence angles
}
\description{
Uses Emission Angle and Incidence Angle to predict DN
}
\examples{
# Numerical
# ---------
minnaert(12.345, 323.3425, .4)

# Vector
# ------

}
